requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
duckduckgo-search==4.1.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
